Refactor: remove core design system from common module
