+++ core/common/build.gradle.kts
@@ -3,7 +3,4 @@ plugins {
 }
 
 android.namespace = "com.waffiq.bazz_movies.core.common"
-
-dependencies {
-  implementation(project(":core:designsystem"))
-}
+dependencies { implementation(libs.androidx.core.ktx) }
