+++ core/favoritewatchlist/src/main/kotlin/com/waffiq/bazz_movies/core/favoritewatchlist/utils/common/Constants.kt
@@ -1,0 +1,8 @@
+package com.waffiq.bazz_movies.core.favoritewatchlist.utils.common
+
+import com.waffiq.bazz_movies.core.designsystem.R.string.movies
+import com.waffiq.bazz_movies.core.designsystem.R.string.tv_series
+
+object Constants {
+  val tabMoviesTvHeadingArray = intArrayOf(movies, tv_series)
+}
