+++ feature/home/src/main/kotlin/com/waffiq/bazz_movies/feature/home/ui/HomeFragment.kt
@@ -9,10 +9,10 @@ import androidx.fragment.app.Fragment
 import androidx.viewpager2.widget.ViewPager2
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
-import com.waffiq.bazz_movies.core.common.utils.Constants.tabHomeHeadingArray
 import com.waffiq.bazz_movies.core.designsystem.R.string.binding_error
 import com.waffiq.bazz_movies.core.designsystem.viewpager.GenericViewPagerAdapter
 import com.waffiq.bazz_movies.feature.home.databinding.FragmentHomeBinding
+import com.waffiq.bazz_movies.feature.home.utils.common.Constants.tabHomeHeadingArray
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
