+++ feature/home/src/main/kotlin/com/waffiq/bazz_movies/feature/home/utils/common/Constants.kt
@@ -1,0 +1,9 @@
+package com.waffiq.bazz_movies.feature.home.utils.common
+
+import com.waffiq.bazz_movies.core.designsystem.R.string.featured
+import com.waffiq.bazz_movies.core.designsystem.R.string.movies
+import com.waffiq.bazz_movies.core.designsystem.R.string.tv_series
+
+object Constants {
+  val tabHomeHeadingArray = intArrayOf(featured, movies, tv_series)
+}
