+++ feature/watchlist/src/main/kotlin/com/waffiq/bazz_movies/feature/watchlist/ui/MyWatchlistFragment.kt
@@ -8,9 +8,9 @@ import androidx.fragment.app.Fragment
 import androidx.viewpager2.widget.ViewPager2
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
-import com.waffiq.bazz_movies.core.common.utils.Constants.tabMoviesTvHeadingArray
 import com.waffiq.bazz_movies.core.designsystem.R.string.binding_error
 import com.waffiq.bazz_movies.core.designsystem.viewpager.GenericViewPagerAdapter
+import com.waffiq.bazz_movies.core.favoritewatchlist.utils.common.Constants.tabMoviesTvHeadingArray
 import com.waffiq.bazz_movies.feature.watchlist.databinding.FragmentMyWatchlistBinding
 import dagger.hilt.android.AndroidEntryPoint
